Created a semi-functional tetrisGame.c
This is the current framework for Tetris. It spawns a random tetromino (Tetris piece) and lowers it over time, while allowing the user to move the piece left, right, or down, then spawns a new piece when it falls down onto a surface. Collision detection works and should be usable for rotation when rotation is implemented.  Line clearing has not been implemented either.

Many of the game objects are generalized, so that interactions work with different proportions, but others need attention and the tetrominoes themselves aren't scalable.

This  version of the code includes a global 3D array to temporarily skirt the development complications of a potentially necessary tetrominoes.h, which would just include that 3D array.

I know checkOverlap() is ugly as heck right now. It carries a lot of weight for the code and hopefully it can be optimized as well.

This version of the code includes a placeholder menu. It also includes test input handling for Windows, which can be enabled in Windows by de-commenting everything below the two "/*Windows:*/" tags and commenting everything below the "/*Linux:*/" tags, as well as de-commenting the body of getLeftInput(). The active piece can then be directed with the WASD keys. The "hard drop" should be functional but it isn't assigned to any return value of getLeftInput right now (because getLeftInput will soon be changed for all games).